I think I might have finished the local routine. Some further checks need to be done to make sure but it looks good so far. Updated the symmetrize_new.py file to also read the g4iwpp-worm component and writes it to the new file which is needed for asymptotics.

+++ src/local_n_point.py
@@ -299,7 +299,7 @@ class LocalNPoint(IHaveMat):
             + (2 * niv,) * num_fermionic_frequency_dimensions
         )
         return LocalNPoint(
-            np.full(shape, value, dtype=np.complex64),
+            np.full(shape, value, dtype=np.complex128),
             num_orbital_dimensions,
             num_bosonic_frequency_dimensions,
             num_fermionic_frequency_dimensions,
